requirements.txt
tensorflow[and-cuda]==2.15.1
tensorflow-recommenders==0.7.3
tensorflow-probability==0.23.0
tensorboard
optuna
ImbalancedLearningRegressionpandas
numpy
seaborn
matplotlib
plotly
wordcloud
